Refactor NavLink into AsideNav (no need for it to exist as separate component - not reusable!)
